apps/requirements.txt
Django>=2.1.6
djangorestframework==3.9.3
django-filter
coreapi
django-rest-auth[with_social]
Markdown
Pygments
django-model-utils
inflection
drf-extensions
PyMySQL
mysqlclient
python-dotenv
django-guardian
djangorestframework-guardian

# inspection
django-silk


# testing
pytest-django
mock
factory-boy
tox

# static code analysis
flake8
flake8-quotes

# others
django-extensions
jupyter

